requirements.txt
jinja2
pillow
pydantic
rich
streamdeck
websockets
yaml
